Rename secureCreatives file and lint
